Fix the playback crash on iOS.
Make sure the size of each buffer pull is less than the size of the
buffer, for if they are the same, the read/write indeces wrap around
and it cannot be determined that the buffer was filled. This results
in an endless refilling of the audio output buffer and a stalling
of the audio worker thread.

That stall in turn prevents the app from creating the audio player,
and so the app would assert upon trying to play.