Fix crash using liblouis on iOS
The default stack size for iOS is 512k, and the qt_main_wrapper()
in qtioseventdispatcher.mm also uses 512k as the default. This is much
too small for liblouis to compile its tables, so we use Qt's
info.plist value to set the stack size to 8MB.

This doesn't crash on the Mac because the default stack size from
the system is 8MB, regardless of what a thread might request.

The proper fix is arguably to get liblouis to use less stack. The
main offender is compileRule() in compileTranslationTable.c.