Create notification listener for app events
For iOS we need a way to catch system notifications so that we can
do the right thing when the app is put into the background or
brought forward. This is especially important when iOS is about to
quit the app in order to avoid an error message about the previous
run not terminating cleanly.