Improve iOS 15 compatibility
Add arm64 libsndfile.a libraries which have been built
for a deployment target of iOS 15. There is likely no real
code change to this, but it prevents a slew of build warnings
that the library is too new for the build deployment target.

Also deal with an assertion that seems to show up only on iOS
15 but is not actually a problem.

And update the state of running MSS on iOS.