Force soundfont control to always be active
This change is for tablet UI. It forces the soundfont control to always
be active, since the mouse version requires a hover before the control
can be activated. That works on mouse systems but not on touches.

This means this change probably fouls up mouse input and doesn't look
right. We need to address the larger issue of how to adapt the UI
to deal with both tablet and mouse UI without near-duplicating
the UI.