Tweak cmake build so Mac side works
Put iOS-only source files out of view of the Mac build. Set up
Qt dir for Mac-side builds.

Somehow OSXAudioDeviceID was turned into a string, whereas for
osxaudiodriver.cpp it clearly needs to be a uint32_t. Not sure
what's going on there, perhaps we are overdue for a TOT merge.